move formatting to new file
